requirements.txt
PyYAML>=6.0
pathlib>=1.0.1 